gitlab: load yadage spec and calculate workflow complexity
closes https://github.com/reanahub/reana-server/issues/368
